node-main/deps/v8/test/inspector/debugger/wasm-conditional-breakpoints-expected.txt

Test conditional breakpoints in wasm.

Running test: test
Instantiating.
Waiting for wasm script.
Got wasm script: wasm://wasm/c5d9f10e
Setting breakpoint at offset 34, condition "false"
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 34
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
Setting breakpoint at offset 41, condition "true"
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 41
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
Setting breakpoint at offset 46, condition "$var0.value==3"
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 46
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
Calling fib(5)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 5 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 4 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 3 (i32)
Script wasm://wasm/c5d9f10e byte offset 46: Wasm opcode 0x10 (kExprCallFunction)
$var0: 3 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 2 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 1 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 2 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 3 (i32)
Script wasm://wasm/c5d9f10e byte offset 46: Wasm opcode 0x10 (kExprCallFunction)
$var0: 3 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 2 (i32)
Script wasm://wasm/c5d9f10e byte offset 41: Wasm opcode 0x0d (kExprBrIf)
$var0: 1 (i32)
fib returned!
